[ALPS07535869] disp: fixed coverity
[Detail]
Fix coverity via explicit casting u64 to avoid overflow

MTK-Commit-Id: f7d581810fd6296e2fd03ea9e7ee48cb17d0a4ae

CR-Id: ALPS07535869
Feature: display_mobile

+++ drivers/gpu/drm/mediatek/mediatek_v2/mtk_drm_crtc.c
@@ -2789,7 +2789,7 @@ _mtk_crtc_cwb_addon_module_connect(
 					src_roi_r;
 				addon_config.addon_wdma_config.wdma_dst_roi =
 					dst_roi_r;
-				addon_config.addon_wdma_config.addr += r_buff_off * Bpp;
+				addon_config.addon_wdma_config.addr += (u64)r_buff_off * (u64)Bpp;
 				/* connect right pipe */
 				mtk_addon_connect_after(crtc, ddp_mode, addon_module,
 							  &addon_config, cmdq_handle);
